fix:idをLongに変更して、日時を適用した

+++ task5/app/src/main/java/com/example/task5/Message.kt
@@ -1,7 +1,9 @@
 package com.example.task5
 
+import java.math.BigInteger
+
 class Message(
-    val id: Int = 0,
+    val id: Long = 0,
     val type: Int = 0,
     val message: String = "",
     val sendTime: String? = ""
